Tweaks to Base
